cdk/pyproject.toml
[tool.poetry]
name = "cdk"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
aws-cdk-lib = "^2.62.1"
constructs = ">=10.0.0,<11.0.0"
black = "^22.12.0"
aws-cdk-aws-lambda-python-alpha = "^2.62.1a0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
